Use double everywhere in mael code

+++ src/mael/teensy/drive.cpp
@@ -7,18 +7,18 @@
 #include "drive.hh"
 
 // Drive speed
-static float drv_x;
-static float drv_y;
-static float drv_w;
+static double drv_x;
+static double drv_y;
+static double drv_w;
 
 // Motor control
-static float dc_angle = 0.0f;
+static double dc_angle = 0.0;
 static int dc_ticks[4] = {};
 static int dc_speed[4] = {};
 
 // PID state
-static float pid_acc[4] = {};
-static float pid_prev[4] = {};
+static double pid_acc[4] = {};
+static double pid_prev[4] = {};
 
 static inline void IncrementEncoderSpeed(int idx, int dir_pin)
 {
@@ -74,13 +74,13 @@ void ProcessDrive()
     // -1/sqrt(2) |  1/sqrt(2) | 1
 
     // DC speed constants
-    static const float kl = 1.0f;
-    static const float kw = 1.0f;
+    static const double kl = 1.0;
+    static const double kw = 1.0;
 
     // PID constants
-    static const float kp = 1.0f;
-    static const float ki = 0.0f;
-    static const float kd = 0.0f;
+    static const double kp = 1.0;
+    static const double ki = 0.0;
+    static const double kd = 0.0;
 
     int ticks[4];
     noInterrupts();
@@ -89,31 +89,31 @@ void ProcessDrive()
     interrupts();
 
     // Eventually we will integrate gyroscope information (Kalman filter)
-    dc_angle += (float)ticks[0] / kw + (float)ticks[1] / kw +
-                (float)ticks[2] / kw + (float)ticks[3] / kw;
+    dc_angle += (double)ticks[0] / kw + (double)ticks[1] / kw +
+                (double)ticks[2] / kw + (double)ticks[3] / kw;
 
     // World coordinates to robot coordinates
-    float self_x = drv_x * cosf(-dc_angle) - drv_y * sinf(-dc_angle);
-    float self_y = drv_x * sinf(-dc_angle) + drv_y * cosf(-dc_angle);
-    float self_w = drv_w;
+    double self_x = drv_x * cosf(-dc_angle) - drv_y * sinf(-dc_angle);
+    double self_y = drv_x * sinf(-dc_angle) + drv_y * cosf(-dc_angle);
+    double self_w = drv_w;
 
     // Compute target speed for all 4 motors
     int target[4] = {};
     {
-        float x = self_x * kl;
-        float y = self_y * kl;
-        float w = self_w * kw;
+        double x = self_x * kl;
+        double y = self_y * kl;
+        double w = self_w * kw;
 
-        target[0] = (int)(x * -0.7071f + y * -0.7071f + w * 1.0f);
-        target[1] = (int)(x *  0.7071f + y * -0.7071f + w * 1.0f);
-        target[2] = (int)(x *  0.7071f + y *  0.7071f + w * 1.0f);
-        target[3] = (int)(x * -0.7071f + y *  0.7071f + w * 1.0f);
+        target[0] = (int)(x * -0.7071 + y * -0.7071 + w * 1.0);
+        target[1] = (int)(x *  0.7071 + y * -0.7071 + w * 1.0);
+        target[2] = (int)(x *  0.7071 + y *  0.7071 + w * 1.0);
+        target[3] = (int)(x * -0.7071 + y *  0.7071 + w * 1.0);
     }
 
     // Run target DC speeds through PID controller
     for (int i = 0; i < 4; i++) {
-        float error = (float)(target[i] - ticks[i]);
-        float delta = error - pid_prev[i];
+        double error = (double)(target[i] - ticks[i]);
+        double delta = error - pid_prev[i];
 
         ticks[i] = 0;
         pid_acc[i] += error;
@@ -129,9 +129,9 @@ void ProcessDrive()
     WriteMotorSpeed(PIN_DC3_DIR, PIN_DC3_PWM, dc_speed[3]);
 }
 
-void SetDriveSpeed(float x, float y, float w)
+void SetDriveSpeed(double x, double y, double w)
 {
     drv_x = x;
     drv_y = y;
-    drv_w = fmodf(-w, 360.0f) * PI / 180.0f;
+    drv_w = fmodf(-w, 360.0) * PI / 180.0;
 }
